Increase buffer version possibilities

+++ filament/backend/src/opengl/OpenGLContext.h
@@ -53,11 +53,11 @@ public:
 
         // If this version number does not match vertexBufferWithObjects->bufferObjectsVersion,
         // then the VAO needs to be updated.
-        uint8_t vertexBufferVersion = 0;
+        uint32_t vertexBufferVersion = 0;
         
         // If this version number does not match indexBuffer->bufferObjectVersion,
         // then the VAO needs to be updated.
-        uint8_t indexBufferVersion = 0;
+        uint32_t indexBufferVersion = 0;
     } gl;
 
     OpenGLContext() noexcept;
